[ELECTRE_TRI] Avoid floating point issues...
THe representation of floating points might be confusing.
Indeed a comparison of two floating points snumber that looks similar
might give an unequality.
e.g. a = 0.00015; b = 0.00015000000001 will be represented both as 0.00015
but when the comparison will be done, it will give an inequality.
This patch add a tolerance on the equality