sat_rmp: update example and fix small error in formulation
There was an issue when a profile wa above all the alternatives on one
criteria since we choose the best performance among the alternatives in
order to initialize the profiles. To fix that, we add a small espsilon
value.

We also remove the profile order algorithm because it is useless. It
just allows to check that the model does not vialoate dominance.